+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Copyright (c) 2021-2024 Claudio Satriano <satriano@ipgp.fr>
 - Filter catalog files on reading using the criteria in the config file
 - Improved time axes in `plot_timespans` and `plot_slip` for short time
   intervals
+- `print_families`: autoset duration units based on the average duration
 
 ## v0.4.1 - 2024-03-11